Fixed scatter plot and saving outputs
There was an issue where the colormap didn't cover the range of values for the expression plots over the UMAP. This was fixed by adjusting the range.

There were also issues with changing the AnnData object to the background corrected values to save the h5ad and feather output files. This was fixed to work with the Logicle normalization step.