Fix channel name issue
The name field of the AnnData object was not being preserved because InfinityMarker channel names were different between files. To correct this test case, the name field will be reset to an empty string when the regression models are built. The names of the backbone channels provided by the user in the FCS file must be the same for them to be preserved. This correction will allow for the non-backbone channels to have different names without breaking the ability to propagate the backbone channel names.